Allow interaction to start piece placements
Began working on the logic and questions that will eventually allow for players to put in coordinates and the place the pieces after checks for possible errors. STILL IN PROGRESS

Changed how marking works to switch from either X 's or 0's (hitting or placing)

Misc changes to remove redundant/repeated questions.